% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{district1950}
\alias{district1950}
\title{Indian district map circa 1956}
\format{A SpatialPolygonsDataFrame object}
\usage{
district1950
}
\description{
Map showing Indian district boundaries... TODO
}
\examples{
data(district1950)
}
\keyword{datasets}

