% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icrisat1950}
\alias{icrisat1950}
\title{Aggregated ICRISAT...}
\format{A data.frame}
\usage{
icrisat1950
}
\description{
Dataset containing various.. TODO
}
\examples{
data(icrisat1950)
}
\keyword{datasets}

