% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indiastat}
\alias{indiastat}
\title{District level agricultural statistics for India}
\format{A data.frame}
\usage{
indiastat
}
\description{
Dataset containing various... TODO
}
\examples{
data(indiastat)
}
\keyword{datasets}

