% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indiastat1950}
\alias{indiastat1950}
\title{Aggregated district level agricultural statistics for India}
\format{A data.frame}
\usage{
indiastat1950
}
\description{
Dataset containing various... TODO
}
\examples{
data(indiastat1950)
}
\keyword{datasets}

